requirements.txt
allure-python-commons==2.13.5
flake8==7.1.1
pytest==8.3.2
requests==2.32.3
selenium==4.24.0
webdriver-manager==4.0.2
